Server - Fixed handlers expect edit

+++ Backend/Serenity/Modules/Comments/Handlers/DeleteCommentCommandHandler.cs
@@ -2,6 +2,7 @@ using System.Security.Claims;
 using AutoMapper;
 using MediatR;
 using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
 using Serenity.Common;
 using Serenity.Database;
 using Serenity.Database.Entities;
@@ -27,19 +28,42 @@ public class DeleteCommentCommandHandler : IRequestHandler<DeleteCommentCommand,
     public async Task<Response> Handle(DeleteCommentCommand command, CancellationToken token)
     {
         var user = await userManager.GetUserAsync(command.Claims);
-        var comment = context.Comments.Where(x => x.Id == command.CommentId && x.UserId == user.Id && x.PostId == command.PostId).First();
+
+        if (user is null)
+        {
+            return new Response
+            {
+                Errors = new()
+                {
+                    new("UserNotFound", "Could not find the user")
+                }
+            };
+        }
+
+        var comment = context.Comments
+            .Where(x => x.Id == command.CommentId && x.UserId == user.Id && x.PostId == command.PostId)
+            .Include(x => x.Post)
+            .Include(x => x.Replies)
+            .First();
 
         if (comment is null)
+        {
             return new()
             {
                 Success = false,
                 Errors = new() { new("CommentNotFound", $"Could not find the comment of Id {command.CommentId}") }
             };
+        }
 
-        foreach (var reply in comment.Replies)
-            context.Comments.Remove(reply);
+        if (comment.Replies is not null)
+        {
+            foreach (var reply in comment.Replies)
+                context.Comments.Remove(reply);
+        }
 
+        comment.Post.Comments.Remove(comment);
         context.Comments.Remove(comment);
+
         var result = context.SaveChanges();
 
         if (result >= 0)
