+++ Backend/Serenity/Modules/Posts/Handlers/DeletePostCommandHandler.cs
@@ -27,19 +27,37 @@ public class DeletePostCommandHandler : IRequestHandler<DeletePostCommand, Respo
     public async Task<Response> Handle(DeletePostCommand command, CancellationToken token)
     {
         var user = await userManager.GetUserAsync(command.Claims);
+
+        if (user is null)
+        {
+            return new CreatePostResponse
+            {
+                Errors = new()
+                {
+                    new("UserNotFound", "Could not find the user")
+                }
+
+            };
+        }
+
         var post = context.Posts.Where(x => x.Id == command.Id && x.UserId == user.Id).First();
 
         if (post is null)
+        {
             return new()
             {
                 Success = false,
                 Errors = new() { new("PostNotFound", $"Could not find the comment of Id {command.Id}") }
             };
-
-        foreach (Comment comment in post.Comments)
-        {
-            context.Comments.Remove(comment);
         }
+
+        if (post.Comments is not null && post.Comments.Count() > 0)
+        {
+            foreach (Comment comment in post.Comments)
+            {
+                context.Comments.Remove(comment);
+            }
+        } 
 
         user.Posts.Remove(post);
         context.Posts.Remove(post);
@@ -47,11 +65,13 @@ public class DeletePostCommandHandler : IRequestHandler<DeletePostCommand, Respo
         var result = context.SaveChanges();
 
         if (result >= 0)
+        {
             return new()
             {
                 Success = true,
                 Errors = null
             };
+        }
 
         return new()
         {
